# openapi.yaml
openapi: 3.1.0
info:
  title: Shoutboard API
  version: 1.0.0
  description: |
    API-first employee recognition group card platform. Shoutboard enables teams to
    create collaborative recognition cards for birthdays, promotions, farewells, and
    other occasions with public shareable links and team contributions.

    ## Key Features
    - **Multi-occasion support**: Birthday, anniversary, promotion, welcome, farewell, custom
    - **Public sharing**: Generate shareable links for public contribution
    - **Team collaboration**: Collect messages from multiple contributors anonymously or identified
    - **Webhook integrations**: Real-time event notifications for board creation, posting, and delivery
    - **API-first design**: Full programmatic access with JWT or API key authentication
    - **Analytics**: Track board engagement, contributor metrics, and team participation

    ## Authentication Methods
    The API supports two authentication methods:

    ### JWT Bearer Token
    Register and login to receive a JWT token valid for 7 days. Include in Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ### API Key
    Create persistent API keys in the dashboard for service-to-service integrations.
    Include in header:
    ```
    X-API-Key: sb_live_<random_string>
    ```
    API keys can be scoped to specific permissions and revoked at any time.

    ## Webhook Events
    Subscribe to real-time events:
...

...

    ## Rate Limiting
...

...
    ```json
...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    Board:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
          description: Unique shareable URL slug
        occasionType:
          type: string
          enum: [BIRTHDAY, ANNIVERSARY, FAREWELL, PROMOTION, WELCOME, CUSTOM]
        status:
          type: string
          enum: [DRAFT, ACTIVE, SENT]
          description: |
            - DRAFT: Being edited, not yet shareable
            - ACTIVE: Public and accepting contributions
            - SENT: Delivery email sent to recipient
        recipientName:
          type: string
        recipientEmail:
          type: string
          format: email
        coverTheme:
          type: string
          description: Design theme (indigo, rose, amber, emerald, etc.)
        postCount:
          type: integer
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
          description: When to automatically send this board
        creator:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
      required: [id, title, slug, occasionType, status, recipientName, createdAt, creator]
    Post:
      type: object
      properties:
        id:
          type: string
        authorName:
          type: string
        contentText:
          type: string
          maxLength: 5000
        mediaUrl:
          type: string
          format: uri
          nullable: true
        gifUrl:
          type: string
          format: uri
          nullable: true
        isAnonymous:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, authorName, contentText, isAnonymous, createdAt]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
...
paths:
...
  /v1/boards:
    get:
      summary: List boards for organization
      operationId: listBoards
      tags: [Boards]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, ACTIVE, SENT]
          description: Filter by board status
        - name: occasionType
          in: query
          schema:
            type: string
          description: Filter by occasion type
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: List of boards
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      boards:
                        type: array
                        items:
                          $ref: '#/components/schemas/Board'
                      pagination:
                        type: object
                        properties:
                          limit:
                            type: integer
                          offset:
                            type: integer
                          total:
                            type: integer
        401:
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create a new board
      operationId: createBoard
      tags: [Boards]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  example: Happy Birthday Sarah!
                occasionType:
                  type: string
                  enum: [BIRTHDAY, ANNIVERSARY, FAREWELL, PROMOTION, WELCOME, CUSTOM]
                recipientName:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: Sarah Johnson
                recipientEmail:
                  type: string
                  format: email
                  example: sarah@example.com
                coverTheme:
                  type: string
                  default: indigo
                  example: rose
                scheduledAt:
                  type: string
                  format: date-time
                  nullable: true
                  description: When to automatically send (optional)
              required: [title, occasionType, recipientName]
      responses:
        201:
          description: Board created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Board'
        422:
          $ref: '#/components/responses/ValidationError'
  /v1/boards/{id}:
    get:
      summary: Get board with posts
      operationId: getBoard
      tags: [Boards]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Board details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Board'
                      - type: object
                        properties:
                          posts:
                            type: array
                            items:
                              $ref: '#/components/schemas/Post'
        404:
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update board details
      operationId: updateBoard
      tags: [Boards]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                recipientName:
                  type: string
                  minLength: 1
                  maxLength: 100
                coverTheme:
                  type: string
                scheduledAt:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        200:
          description: Board updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Board'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete a board
      operationId: deleteBoard
      tags: [Boards]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Board deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
  /v1/boards/{id}/send:
    post:
      summary: Send board to recipient
      operationId: sendBoard
      tags: [Boards]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      description: |
        Sets board status to SENT and sends delivery email if recipient email provided.
        Triggers webhook events.
      responses:
        200:
          description: Board sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Board'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
  /v1/boards/{id}/schedule:
    post:
      summary: Schedule board to send later
      operationId: scheduleBoard
      tags: [Boards]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduledAt:
                  type: string
                  format: date-time
                  example: '2024-04-15T10:00:00Z'
              required: [scheduledAt]
      responses:
        200:
          description: Board scheduled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Board'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
  /v1/boards/{slug}/public:
    get:
      summary: Get public board view (no auth required)
      operationId: getPublicBoard
      tags: [Boards]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Board slug (URL-safe identifier)
      description: |
        Publicly accessible board view. Only works for boards with status=SENT.
        Records anonymous view count (hashed IP).
      responses:
        200:
          description: Public board view
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      title:
                        type: string
                      recipientName:
                        type: string
                      occasionType:
                        type: string
                      coverTheme:
                        type: string
                      posts:
                        type: array
                        items:
                          $ref: '#/components/schemas/Post'
                      viewCount:
                        type: integer
                      createdAt:
                        type: string
                        format: date-time
                      creator:
                        type: object
        404:
          description: Board not found or not public
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Post Endpoints
...
components:
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Board not found
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Field validation failed